requirements.txt
Flask
requests
python-dotenv
sympy
pytest
Flask-WTF
WTForms
google-generativeai